New Generation of Application
This change introduces a new generation of the Tanzu Java Web
Application.  It's now based on a modern version of Spring Boot, built
with Gradle, and uses a non-blocking web server.  The application has
also been updated with a new UI that exposes more information and
collects that information in a more structured way.

Signed-off-by: Ben Hale <ben.hale@broadcom.com>