Remove Data Caching
Previously, the application was caching all of the general and cloud
information upon first request to improve performance. This meant that
if any change was made to the Space's configuration, e.g. changing
egress rules, the application would not reflect those changes until it
was restarted. This change removes the caching and instead fetches the
general and cloud info on every request.

Signed-off-by: Ben Hale <ben.hale@broadcom.com>