Carvel Artifacts
This change introduces the creation and publication of Carvel artifacts.
These artifacts are built using the Carvel toolchain and are intended to
be used with Kubernetes.

A Makefile is added to the project with three
new commands that create the Package (including building the image with
Cloud Native Buildpacks), the Package Repository, and printing the
Kubernetes manifests required to install the application.

A GitHub Workflow is added to publish the Carvel Package on every commit
to main.

Signed-off-by: Ben Hale <ben.hale@broadcom.com>